fix(huya-extractor): change rsp presenterUid type from Int to Long

+++ platforms/src/main/kotlin/github/hua0512/plugins/huya/danmu/msg/req/HuyaGetTokenResp.kt
@@ -3,7 +3,7 @@
  *
  * Stream-rec  https://github.com/hua0512/stream-rec
  *
- * Copyright (c) 2024 hua0512 (https://github.com/hua0512)
+ * Copyright (c) 2025 hua0512 (https://github.com/hua0512)
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -35,7 +35,7 @@ data class HuyaGetTokenResp(
   var url: String = "",
   var cdnType: String = "",
   var streamName: String = "",
-  var presenterUid: Int = 0,
+  var presenterUid: Long = 0,
   var antiCode: String = "",
   var sTime: String = "",
   var flvAntiCode: String = "",
